+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes for the LINZ BDE schema are documented in this file.
 
 ## 1.11.0dev - 2020-MM-DD
 ### Added
+- Landonline 3.22 support:
+  - New `crs_ttl_inst_protect` table
 ### Enhanced
 ### Changed